Cere check-matching updates the graph
After calling cere check-matching on region(s), the graph is updated.
This remove the update code redundancy in selector. This also allow to call
cere report after cere check-matching and visualize the matching result.
Documentation is also updated accordingly.
